+++ src/jvmMain/kotlin/ui/opentrades/model/OpenTradesEvent.kt
@@ -4,9 +4,9 @@ internal sealed class OpenTradesEvent {
 
     object AddTrade : OpenTradesEvent()
 
-    data class EditTrade(val id: Int) : OpenTradesEvent()
+    data class EditTrade(val id: Long) : OpenTradesEvent()
 
-    data class CloseTrade(val id: Int) : OpenTradesEvent()
+    data class CloseTrade(val id: Long) : OpenTradesEvent()
 
-    data class DeleteTrade(val id: Int) : OpenTradesEvent()
+    data class DeleteTrade(val id: Long) : OpenTradesEvent()
 }
